refactor : 피드백 대응
