+++ app/src/main/java/camp/nextstep/edu/github/MainActivity.kt
@@ -13,7 +13,6 @@ import androidx.lifecycle.repeatOnLifecycle
 import camp.nextstep.edu.github.adapter.RepositoryAdapter
 import camp.nextstep.edu.github.databinding.ActivityMainBinding
 import camp.nextstep.edu.github.domain.GithubRepository
-import camp.nextstep.edu.github.domain.UiState
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 
@@ -54,9 +53,9 @@ class MainActivity : AppCompatActivity() {
         lifecycleScope.launch {
             lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED) {
                 mainViewModel.repositories.collect { uiState: UiState<List<GithubRepository>> ->
-                    Log.e("TestTest", "observeData: $uiState")
+                    binding.progressBar.isVisible = uiState is UiState.Loading
+                    binding.errorGroup.isVisible = uiState is UiState.Error
                     when (uiState) {
-                        is UiState.Loading -> showLoading()
                         is UiState.Success -> handleSuccess(uiState.data)
                         is UiState.Error -> handleError(uiState.throwable)
                     }
@@ -65,23 +64,11 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun showLoading() {
-        binding.progressBar.isVisible = true
-    }
-
     private fun handleSuccess(data: List<GithubRepository>?) {
-        binding.progressBar.isVisible = false
-        binding.errorGroup.isVisible = false
-
-        binding.repositoryRecyclerView.isVisible = true
-
         repositoryAdapter.submitList(data)
     }
 
     private fun handleError(throwable: Throwable?) {
-        binding.errorGroup.isVisible = true
-        binding.progressBar.isVisible = false
-
         Toast.makeText(applicationContext, "$throwable", Toast.LENGTH_SHORT).show()
     }
