+++ app/src/main/java/camp/nextstep/edu/github/MainViewModel.kt
@@ -1,15 +1,13 @@
 package camp.nextstep.edu.github
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import camp.nextstep.edu.github.domain.GithubRepository
 import camp.nextstep.edu.github.domain.NetworkRepository
-import camp.nextstep.edu.github.domain.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -22,12 +20,10 @@ class MainViewModel @Inject constructor(
     val repositories = _repositories.asStateFlow()
 
     fun fetchRepositories() = viewModelScope.launch {
-        networkRepository.getGithubRepositories().collect { uiState: UiState<List<GithubRepository>> ->
-            when (uiState) {
-                is UiState.Success -> _repositories.emit(uiState)
-                is UiState.Error -> _repositories.emit(uiState)
-                is UiState.Loading -> _repositories.emit(uiState)
-            }
+        if (_repositories.value !is UiState.Success) {
+            networkRepository.getGithubRepositories()
+                .onSuccess { _repositories.emit(UiState.Success(it)) }
+                .onFailure { _repositories.emit(UiState.Error(it)) }
         }
     }
