+++ app/src/main/res/layout/activity_main.xml
@@ -21,7 +21,6 @@
             android:id="@+id/repositoryRecyclerView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:visibility="gone"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
 
         <androidx.constraintlayout.widget.Group
