--- domain/src/main/java/camp/nextstep/edu/github/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package camp.nextstep.edu.github.domain
-
-class MyClass {
-}
